ex 4.21 completed

+++ part-4/blog/controllers/blogs.js
@@ -3,16 +3,6 @@ const Blog = require("../models/blog");
 const User = require("../models/user");
 const jwt = require("jsonwebtoken");
 
-// const extractToken = function (request) {
-//   const authorization = request.get("authorization");
-
-//   if (authorization && authorization.startsWith("Bearer ")) {
-//     return authorization.replace("Bearer ", "");
-//   }
-
-//   return null;
-// };
-
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
 
@@ -20,20 +10,20 @@ blogsRouter.get("/", async (request, response) => {
 });
 
 blogsRouter.post("/", async (request, response, next) => {
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-
-  if (!decodedToken.id) {
-    response.status(401).send({ error: "token invalid" });
-  }
-
-  const user = await User.findById(decodedToken.id);
-
-  const payload = {
-    ...request.body,
-    user: user._id,
-  };
-
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).send({ error: "token invalid" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
+    const payload = {
+      ...request.body,
+      user: user._id,
+    };
+
     const blog = new Blog(payload);
 
     const savedBlog = await blog.save();
@@ -65,7 +55,26 @@ blogsRouter.put("/:id", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).send({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (!blog) {
+      return response.status(404).send({ error: "blog not found" });
+    }
+    const user = await User.findById(decodedToken.id);
+
+    if (!(blog.user.toString() === user._id.toString())) {
+      return response.status(401).send({ error: "user not allowed" });
+    }
+
     await Blog.findByIdAndDelete(request.params.id);
+    user.blogs = user.blogs.filter((blog) => blog.id !== request.params.id);
+    await user.save();
     response.status(204).end();
   } catch (error) {
     next(error);
